fix(analysis-quant): satisfy mypy and ruff on new modules
Cast CAGR return for strict mypy and shorten walk-forward assertion line.

+++ quantpilot/backtest/metrics.py
@@ -70,7 +70,7 @@ def cagr(equity: float, start: date, end: date) -> float:
     years = max((end - start).days / 365.25, 1 / 365.25)
     if equity <= 0:
         return -1.0
-    return (equity ** (1 / years)) - 1.0
+    return float((equity ** (1 / years)) - 1.0)
 
 
 def profit_factor(trade_pnls: list[float]) -> float:
